subo ejercicios

+++ 02_Variables/ejercicios/T02Ejercicio04.java
@@ -6,7 +6,7 @@
 
 public class T02Ejercicio04 {
   public static void main(String[] args) {
-    double euros = 7.00;
+    float euros = 7.00f;
     int pesetas = ((int) Math.round(euros * 166.386));
     System.out.println(pesetas);
     System.out.println(euros * 166.386);
